+++ dataland-frontend/src/components/generics/TheNewFooter.vue
