+++ dataland-frontend/src/components/resources/apiKey/UserRolesBadges.vue
