+++ dataland-frontend/src/components/resources/newLandingPage/TheBrands.vue
