+++ dataland-frontend/src/components/resources/newLandingPage/TheHowItWorks.vue
