+++ dataland-frontend/src/components/resources/newLandingPage/TheQuotes.vue
