+++ dataland-frontend/src/components/resources/newLandingPage/TheStruggle.vue
