+++ dataland-frontend/tests/component/components/general/ApiKeyCard.cy.ts
@@ -1,4 +1,4 @@
-import ApiKeyCard from "@/components/general/apiKey/ApiKeyCard.vue";
+import ApiKeyCard from "@/components/resources/apiKey/ApiKeyCard.vue";
 import { KEYCLOAK_ROLE_ADMIN, KEYCLOAK_ROLE_USER } from "@/utils/KeycloakUtils";
 
 describe("Component test for ApiKeyCard", () => {
