+++ dataland-frontend/tests/component/components/resources/TheFooter.cy.ts
@@ -1,4 +1,4 @@
-import TheFooter from "@/components/general/TheFooter.vue";
+import TheFooter from "@/components/generics/TheFooter.vue";
 import { mount } from "cypress/vue";
 import { checkImage } from "@ct/testUtils/ExistenceChecks";
